requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.0
sqlmodel==0.0.22
python-docx==1.1.2
openai==1.51.2
httpx==0.27.2
pytest==8.3.3
ruff==0.6.9
black==24.8.0
